apple's unsigned long is 64 bit

+++ Src/Include/GpuScene.h
@@ -27,7 +27,7 @@ struct AAPLTextureData
 
 struct AAPLMeshData
 {
-  unsigned long _vertexCount,_indexCount,_indexType,_chunkCount,_meshCount,_opaqueChunkCount,_opaqueMeshCount,_alphaMaskedChunkCount,_alphaMaskedMeshCount,_transparentChunkCount,_transparentMeshCount,_materialCount;
+  unsigned long long _vertexCount,_indexCount,_indexType,_chunkCount,_meshCount,_opaqueChunkCount,_opaqueMeshCount,_alphaMaskedChunkCount,_alphaMaskedMeshCount,_transparentChunkCount,_transparentMeshCount,_materialCount;
   void * _vertexData;
   void * _normalData;
   void * _tangentData;
